requirements.txt
pre-commit
paho-mqtt
